[Build] Move common folder into app and establish symbolic link in ext
Move common folder into `diagnostics-app/src`.

Add a symbolic link to common folder in `diagnostics-extension/src`.

This will allow the ts compilers in both directory to compile common
files correctly. This will also enable linter/formatter for the common
projects.

BUG: b/301888904
