+++ diagnostics-app/src/app/diagnostics/diagnostics-card/diagnostics-card.component.spec.ts
@@ -2,7 +2,7 @@ import {ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {DiagnosticsCardComponent} from './diagnostics-card.component';
 
-import {RoutineType} from '@common/telemetry-extension-types';
+import {RoutineType} from 'common/telemetry-extension-types';
 
 describe('DiagnosticsCardComponent', () => {
   let component: DiagnosticsCardComponent;
