+++ diagnostics-app/src/app/diagnostics/diagnostics-dashboard/diagnostics-dashboard.component.ts
@@ -13,9 +13,9 @@ import {Component, OnInit} from '@angular/core';
 import {
   GetAvailableRoutinesResponse,
   RoutineType,
-} from '@common/telemetry-extension-types';
-import {DiagnosticsService} from 'src/app/core/services/diagnostics.service';
-import {VISIBLE_DIAGNOSTICS_CARDS} from 'src/config/config';
+} from 'common/telemetry-extension-types';
+import {DiagnosticsService} from 'app/core/services/diagnostics.service';
+import {VISIBLE_DIAGNOSTICS_CARDS} from 'config/config';
 
 @Component({
   selector: 'app-diagnostics-dashboard',
