+++ diagnostics-app/src/app/events/events-card/events-card.component.spec.ts
@@ -1,8 +1,8 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import {EventCategory} from '@common/telemetry-extension-types';
+import {EventCategory} from 'common/telemetry-extension-types';
 import {EventsCardComponent} from './events-card.component';
-import {EventsService} from 'src/app/core/services/events.service';
+import {EventsService} from 'app/core/services/events.service';
 
 describe('EventsCardComponent', () => {
   let component: EventsCardComponent;
