+++ diagnostics-app/src/app/events/events-card/events-card.component.ts
@@ -8,8 +8,8 @@
  */
 
 import {ChangeDetectorRef, Component, Input, OnInit} from '@angular/core';
-import {EventCategory} from '@common/telemetry-extension-types';
-import {EventsService} from 'src/app/core/services/events.service';
+import {EventCategory} from 'common/telemetry-extension-types';
+import {EventsService} from 'app/core/services/events.service';
 
 enum EventsCardState {
   LISTENING = 'listening',
