+++ diagnostics-app/src/app/layout/header/header.component.ts
@@ -8,9 +8,9 @@
  */
 
 import {Component, EventEmitter, Output} from '@angular/core';
-import {Theme} from 'src/app/core/enums/global.enums';
-import {ThemeService} from 'src/app/core/services/theme.service';
-import {APP_NAME} from 'src/config/config';
+import {Theme} from 'app/core/enums/global.enums';
+import {ThemeService} from 'app/core/services/theme.service';
+import {APP_NAME} from 'config/config';
 
 @Component({
   selector: 'app-header',
