+++ diagnostics-app/src/app/telemetry/telemetry-card/telemetry-card.component.ts
@@ -9,9 +9,9 @@
 
 import {Component, Input, OnInit, OnDestroy} from '@angular/core';
 
-import {TelemetryInfoUnion} from '@common/message';
-import {TelemetryService} from 'src/app/core/services/telemetry.service';
-import {TelemetryInfoType, ResponseErrorInfoMessage} from '@common/message';
+import {TelemetryInfoUnion} from 'common/message';
+import {TelemetryService} from 'app/core/services/telemetry.service';
+import {TelemetryInfoType, ResponseErrorInfoMessage} from 'common/message';
 import {
   defaultTelemetryRefreshInterval,
   refreshIntervals,
