+++ diagnostics-app/src/common/message.ts
