+++ diagnostics-app/src/common/telemetry-extension-types.ts
