+++ diagnostics-app/src/config/config.ts
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import {EventCategory, RoutineType} from '@common/telemetry-extension-types';
-import {TelemetryInfoType} from '@common/message';
+import {EventCategory, RoutineType} from 'common/telemetry-extension-types';
+import {TelemetryInfoType} from 'common/message';
 
 export const APP_NAME = 'telemetry-extension-types Reference App';
