+++ diagnostics-extension/src/__tests__/services/telemetry.spec.ts
@@ -11,7 +11,7 @@ import {
   TelemetryServiceProvider,
 } from '../../services/telemetry';
 import * as fakeData from '../../services/fake_telemetry.data';
-import {TelemetryInfoUnion} from '@common/message';
+import {TelemetryInfoUnion} from '../../common/message';
 
 describe('should return instance of FakeTelemetryService', () => {
   const telemetryService: TelemetryService =
