+++ diagnostics-extension/src/common
@@ -1,0 +1,1 @@
+../../diagnostics-app/src/common
