+++ diagnostics-extension/webpack.common.cjs
@@ -14,7 +14,7 @@ module.exports = {
             {
               test : /\.ts$/,
               use : 'ts-loader',
-              include : [path.resolve(__dirname, 'src')],
+              exclude : [path.resolve(__dirname, 'node_modules')],
             },
           ],
   },
